requirements.txt
requests
pandas
python-dateutil
python-dotenv
pyarrow
zstandard
brotli
streamlit
matplotlib
plotly
